Crate cargo_crev_reviews

This module contains the boilerplate to parse and match URI and POST rpc. The real code for methods is in methods_mod.rs The real content of “static files” is in the module files_mod.rs

Modules

cargo_mod

common_mod

crev_mod

files_mod

response_get_mod

response_post_mod

rpc_methods_mod

stdio_input_password_mod

Structs

CLEAR_ALL
ansi clear all

CLEAR_LINE
ansi clear line

CREV_UNLOCKED
mutable static, because it is hard to pass variables around with async closures

GREEN
ansi color

LOCAL

RED
ansi color

RESET
ansi reset color

UNHIDE_CURSOR
ansi unhide cursor

YELLOW
ansi color

Functions

start_web_server

unlock_crev_id_interactively
unlock crev_id interactively