Module cargo_crev_reviews::crev_mod

functions around the crev data and lib

Structs

CREV_LOCAL

CREV_UNLOCKED

mutable static, because it is hard to pass variables around with async closures

PackageSegment

ProofCrevForReview

only the fields I care about for Review

ReviewSegment

TrustedPublisher

TrustedPublisherDataFile

VcsInfoJson

Enums

VcsInfoJsonGit

Constants

VCS_INFO_JSON_FILE

Functions

_verify_proof

cargo_min_ignore_list

crev_crate_versions

get all versions for one crate

crev_edit_or_new_review

edit or new

crev_edit_review

edit my reviews, find it in list

crev_list_my_reviews

list my reviews

crev_new_version

list the user-selected crate+version then find the newest version in cargo registry index and change the version field all the rest stays the same

crev_publish

crev_save_review

create save review proof

crev_sort_list_by_name_version

delete_review_proofs

remove old proofs, so the new review proof will be unique

is_trusted_publisher

load_trusted_publishers_json

path_to_trusted_publishers_json

proof_crev_files_paths

published_by_login

check if it is already in the cache or GET from crates.io API and store in cache

rating_or_version

if exists my review for crate version returns the rating if exists any review for this crate returns version number else return empty string

rating_parse

rating_to_string

reviews_dir

unlock_crev_id_interactively

unlock crev_id interactively

vcs_info_to_revision_string

verify_project

verify_project should return some data quickly, but in the background start to fill the db_version for all these crates. So the next time we have more complete data

verify_sort_list_by_name_version