Module cargo_crev_reviews::db_sled_mod

db_sled is a database that works in memory, but is persisted on disk it looks very fast. It must be used uniquely by only one process! Be careful when changing the shape of the data: then there must exist a migration program to upgrade the data store

Modules

db_crate_mod

cached data for crates for needed crates it reads (sync) from crates.io in the background

db_publisher_mod

persistent storage of “my trusted publishers urls” be careful if the struct changes, then there must be an migration upgrade of the data url is better than login, because is unique

db_review_mod

cached data for “my reviews” reading from yaml files is slow, therefore it reads (sync) from “my reviews” in the background only one review per crate_name_version is permitted

db_verify_mod

cached data for the command “verify” the command can be slow therefore it reads (sync) the result of the command in the background

db_version_mod

cached data for crate versions for needed crates it reads (sync) from crates.io api in the background

db_yanked_mod

cached list of yanked versions in the sled database so the reading should be faster than reading from cargo registry index this is not immutable like other information from crates.io and it needs regular updates it can read the data from cargo registry.

Structs

DB_SLED

“sled” db stays open all the time of the program running. this program on start checks if there is an instance already running, so to guarantee only one process access the db files.

DB_SLED_METADATA

I will store the version here, so I can use it for migration upgrades

POOL

3 threads to download in parallel

Functions

db_sled_migration_update

if the data structure changes then it must be updated by a migration

download_in_background_crate_versions

to avoid double or triple call to crates.io for the same crate

sync_in_background_reviews

working with crev data looks slow. I will sync in background with fast sled database and work from there.

sync_in_background_verify

working with crev verify looks slow. I will sync in background with fast sled database and work from there.

sync_in_background_yanked

from cargo registry read all versions that are in db_version only the yanked store in db_yanked

upgrade_to_2021_1228_1255

upgrade_to_2021_1228_1528