Module cargo_crev_reviews::response_get_mod

response for the GET method. Mainly the static files, that are embedded in Rust code.

Structs

ResponseWithBytes

Enums

Cache

Status

Functions

default_mime

parse_get_uri_and_response_file

GET is used only to request files Files are stored in functions in the auto_generated_files_mod.rs module there is an automation task to copy files from web_server_folder to the module

response_404_not_found

response_file_base64

response_file_text