Module cargo_crev_reviews::srv_methods_mod::srv_review_mod

rpc methods prepare the data to respond the POST rpc requests

Functions

request_review_list

srv_review_delete

srv_review_edit

srv_review_edit_or_new

edit the review or copy the last review to create a new review

srv_review_new

srv_review_new_version

srv_review_open_source_code

The source code of the dependency crate is in the cargo registry src folder But it must not be opened with a code editor, because the intellisense server will alter the files: add the target folder and Cargo.lock file. This is why I unpack the .crate file into a temp directory first. So I avoid also copying unclean crate folders. If the cache file does not exist in cargo registry, I will download it in another temp folder.

srv_review_publish

srv_review_save

srv_reviews_list

maybe add filter for one crate_name