Module cargo_crev_reviews::srv_methods_mod::srv_utils_mod

rpc methods prepare the data to respond the POST rpc requests

Functions

srv_cargo_tree_project

srv_config_edit

srv_config_save

srv_correct_digest

correct digest to all reviews where needed

srv_fix_missing_src_folder

fix the scr folder for crates that have our review

srv_list_unclean_crates

list crates to delete manually. The next cargo check or cargo build will unzip them from the cache folder.

srv_update_registry_index

srv_version_list

list of all versions for one crate: from registry index with data from src cached and my_reviews