Module cargo_crev_reviews::utils_mod

Functions

from_crev_to_item
convert ProofCrevForReview into ReviewItemData

join_crate_version

ns_elapsed
returns the elapsed nanoseconds

ns_print_ms
print elapsed time in milliseconds and returns the new now in nanoseconds

ns_print_ns
print elapsed time in nanoseconds and returns the new now in nanoseconds

ns_start
returns the now in nanoseconds

split_crate_version