Module cargo_crev_reviews::cargo_registry_mod

functions around the cargo registry d ~.cargo\registry\index\github.com-1ecc6299db9ec823\an\yh\anyhow
~.cargo\registry\src\github.com-1ecc6299db9ec823\anyhow-1.0.37\

Structs

CARGO_REGISTRY_CACHE_DIR

CARGO_REGISTRY_SRC_DIR

FileMetadata

Functions

cargo_crev_reviews_cache_file_for_crate

cargo_crev_reviews .crate file for a crate in the temp cache folder

cargo_crev_reviews_src_dir_for_crate

cargo_registry_cache_file_for_crate

cargo registry cache file .crate for a crate

cargo_registry_src_dir_for_crate

cargo registry src directory for a crate

download_crate_from_crate_io

download url from crates.io into cargo_crev_reviews cache temp folder URL like: https://crates.io/api/v1/crates/block-cipher-trait/0.6.0/download it works also for yanked crates. This downloads are immutable.

if_unclean_add_to_ret_list

compare if the folder is unclean push command to delete unclean folder into string ret_list

info_for_one_crate

info of all versions for one crate

is_crate_clean

compare registry cache to registry src if they are equal

list_unclean_crates

compare the files in src and cache folders of cargo registry return a list of unclean folders to manually delete Cargo will download and unpack the next time it needs this crate. For cargo_crev_reviews I will download and unpack in a temp folder.

max_version_from_registry_index

max version from registry index cache

open_cargo_index

unpack_from_targz_to_folder

unpack the .crate file from the cache to any folder

update_registry_index

fetch/pull the registry index from github

yanked_for_one_crate

list only yanked versions for one crate

yanked_for_version

yanked